# docs/swagger.yaml
basePath: /api/v1/wallet
definitions:
  resp.AccountDetailListResp:
    properties:
      items:
        description: 帐号详情
        items:
          $ref: '#/definitions/resp.AccountDetailResp'
        type: array
      page:
        $ref: '#/definitions/resp.BasePageResp'
        description: 分页
        type: object
    type: object
  resp.AccountDetailResp:
    properties:
      currency_id:
        description: 币种id
        type: integer
      decimals:
        description: 小数点位数
        type: integer
      income:
        description: 本期收入
        type: number
      name:
        description: 名称
        type: string
      spend:
        description: 上期支出
        type: number
      symbol:
        description: 币种标识
        type: string
      type:
        description: 明细类型  账单业务类型 1充值 2提币 3节点分红 4算力释放 5转入 6升级
        type: integer
      updated_at:
        description: 时间
        type: string
    type: object
  resp.AccountInfoResp:
    properties:
      balance:
        description: 余额
        type: number
      currency_id:
        description: 币种id
        type: integer
      decimals:
        description: 小数点位数
        type: integer
      name:
        description: 名称
        type: string
      symbol:
        description: 币种标识
        type: string
    type: object
...
  resp.BasePageResp:
    properties:
      count:
        description: 数量
        type: integer
      currentPage:
        description: 当前页数
        type: integer
      pageSize:
        description: 每页数据条数
        type: integer
      totalPage:
        description: 总页数
        type: integer
    type: object
...
  resp.ReadBlockChainListResp:
    properties:
      chain_code:
        description: 链标识
        type: string
      id:
        description: 链id
        type: integer
      name:
        description: 链名称
        type: string
    type: object
...
  resp.ReadOrderSymbolByChainResp:
    properties:
      chain_code:
        description: 链标识
        type: string
      coin_id:
        description: 代币id
        type: integer
      id:
        description: 链id
        type: integer
      name:
        description: 链名称
        type: string
    type: object
...
info:
  contact: {}
  description: 测试
  license: {}
  title: 测试
  version: 0.0.1
paths:
  /account/change:
    post:
      description: 兑换代币
      parameters:
      - description: 代币id
        in: formData
        name: currency_id
        required: true
        type: integer
      - description: 兑换id
        in: formData
        name: change_id
        required: true
        type: integer
      - description: 转账金额
        in: formData
        name: money
        required: true
        type: number
      - description: 转账比例
        in: formData
        name: ratio
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 兑换代币接口
      tags:
      - Account 钱包帐号
  /account/detail:
    get:
      description: 钱包明
      parameters:
      - description: 页面条数
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 页数
        in: query
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.AccountDetailListResp'
      security:
      - ApiKeyAuth: []
      summary: 钱包明接口
      tags:
      - Account 钱包帐号
  /account/info:
    get:
      description: 钱包详情
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.AccountInfoResp'
      security:
      - ApiKeyAuth: []
      summary: 钱包详情接口
      tags:
      - Account 钱包帐号
  /account/share/bonus:
    post:
      description: 节点分红入账
      parameters:
      - description: 金额
        in: formData
        name: money
        required: true
        type: integer
      - description: 币种标示
        in: formData
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 节点分红入账接口
      tags:
      - Account 钱包帐号
  /account/tfor/info:
    get:
      description: 钱包TFOR详情
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.AccountInfoResp'
      security:
      - ApiKeyAuth: []
      summary: 钱包TFOR详情接口
      tags:
      - Account 钱包帐号
  /account/transfer:
    post:
      description: 转账动态usdd
      parameters:
      - description: 转账金额
        in: formData
        name: money
        required: true
        type: number
      - description: 币种标识
        in: formData
        name: symbol
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 转账动态usdd接口
      tags:
      - Account 钱包帐号
  /account/withdrawal:
    post:
      description: 提现货币
      parameters:
      - description: 金额
        in: formData
        name: money
        required: true
        type: number
      - description: 代币id
        in: formData
        name: coin_id
        required: true
        type: integer
      - description: 币种id
        in: formData
        name: currency_id
        required: true
        type: integer
      - description: 提现地址id
        in: formData
        name: withdrawal_addr_id
        required: true
        type: integer
      - description: 支付密码
        in: formData
        name: password
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 提现货币接口
      tags:
      - Account 钱包帐号
...
  /chain/add:
    post:
      description: 增加链
      parameters:
      - description: 链标识
        in: formData
        name: chain_code
        required: true
        type: integer
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 增加链接口
      tags:
      - BlockChain 链
  /chain/list:
    get:
      description: 读取链列表
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadBlockChainListResp'
      security:
      - ApiKeyAuth: []
      summary: 读取链列表接口
      tags:
      - BlockChain 链
  /chain/remove:
    post:
      description: 删除链
      parameters:
      - description: 链id
        in: query
        name: chain_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 删除链接口
      tags:
      - BlockChain 链
  /chain/symbol:
    get:
      description: 根据symbol获取链
      parameters:
      - description: 链标识
        in: formData
        name: symbol
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ReadOrderSymbolByChainResp'
      security:
      - ApiKeyAuth: []
      summary: 根据symbol获取链接口
      tags:
      - BlockChain 链
  /chain/update:
    post:
      description: 更新链
      parameters:
      - description: 链id
        in: formData
        name: chain_id
        required: true
        type: integer
      - description: 链标识
        in: formData
        name: chain_code
        required: true
        type: string
      - description: 名称
        in: formData
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200": {}
      security:
      - ApiKeyAuth: []
      summary: 更新链接口
      tags:
      - BlockChain 链
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"